Convert Continuous mode to use shorts natively
In the interest of speed, reducing the number of buffer swaps and type
conversions in continuous view to match the trigger view. Types are now
converted once before heading to the datastore for their appropriate
type of view.

+++ src/com/backyardbrains/audio/AudioService.java
@@ -62,7 +62,7 @@ public class AudioServiceBinder extends Binder {
 	 * 
 	 * @return a ordinate-corrected version of the audio buffer
 	 */
-	public byte[] getAudioBuffer() {
+	public short[] getAudioBuffer() {
 		return audioBuffer.getArray();
 	}
